+++ components/bridge/api/src/main/java/com/flipperdevices/bridge/api/utils/Constants.kt
@@ -22,6 +22,10 @@ object Constants {
         majorVersion = 0,
         minorVersion = 11
     )
+    val API_SUPPORTED_LOAD_FAP = SemVer(
+        majorVersion = 0,
+        minorVersion = 18
+    )
     val API_SUPPORTED_GET_REQUEST = API_SUPPORTED_FLIPPER_ERROR
     val API_SUPPORTED_VERSION = SemVer(majorVersion = 0, minorVersion = 3)
     const val LAGS_FLIPPER_DETECT_TIMEOUT_MS = 5 * 1000L // 5 seconds