+++ components/faphub/fapscreen/impl/src/main/java/com/flipperdevices/faphub/fapscreen/impl/viewmodel/FapScreenViewModel.kt
@@ -125,7 +125,7 @@ class FapScreenViewModel @VMInject constructor(
             fapItem
         )
 
-        FapState.Installed,
+        is FapState.Installed,
         is FapState.ReadyToUpdate -> FapDetailedControlState.Installed(
             fapItem
         )