+++ components/faphub/installation/button/impl/src/main/java/com/flipperdevices/faphub/installation/button/impl/composable/ComposableFapButtons.kt
@@ -85,6 +85,26 @@ fun ComposableFapInstalledButton(
 }
 
 @Composable
+fun ComposableFapOpenButton(
+    fapButtonSize: FapButtonSize,
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit,
+) {
+    val text = when (fapButtonSize) {
+        FapButtonSize.COMPACTED -> stringResource(R.string.faphub_installation_open)
+        FapButtonSize.LARGE -> stringResource(R.string.faphub_installation_open_long)
+    }
+
+    ComposableFlipperButton(
+        modifier = modifier,
+        text = text,
+        color = LocalPallet.current.accentSecond,
+        fapButtonSize = fapButtonSize,
+        onClick = onClick
+    )
+}
+
+@Composable
 fun ComposableFapUpdateButton(
     fapButtonSize: FapButtonSize,
     modifier: Modifier = Modifier,